# app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/student/{student_id}/history": {
            "get": {
                "summary": "Get Student",
                "operationId": "get_student_student__student_id__history_get",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Student Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HDStudent"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/student/{student_id}/progress": {
            "get": {
                "summary": "Get Student Progress",
                "operationId": "get_student_progress_student__student_id__progress_get",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Student Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/students/email/{email}": {
            "get": {
                "summary": "Get Student Profile By Email",
                "operationId": "get_student_profile_by_email_students_email__email__get",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/XProfile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/student/{student_id}/progress2": {
            "get": {
                "summary": "Get Student Progress2",
                "operationId": "get_student_progress2_student__student_id__progress2_get",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Student Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RAResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/student/{student_name}": {
            "get": {
                "summary": "Read Student",
                "operationId": "read_student_student__student_name__get",
                "parameters": [
                    {
                        "name": "student_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Student Name"
                        }
                    },
                    {
                        "name": "benchmarks",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "title": "Benchmarks"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "HDStudent": {
                "properties": {
                    "people_id": {
                        "type": "integer",
                        "title": "People Id"
                    },
                    "program": {
                        "type": "string",
                        "title": "Program"
                    },
                    "degree": {
                        "type": "string",
                        "title": "Degree"
                    },
                    "curriculum": {
                        "type": "string",
                        "title": "Curriculum"
                    },
                    "college": {
                        "type": "string",
                        "title": "College"
                    },
                    "department": {
                        "type": "string",
                        "title": "Department"
                    },
                    "years": {
                        "items": {
                            "$ref": "#/components/schemas/HDYear"
                        },
                        "type": "array",
                        "title": "Years",
                        "default": []
                    },
                    "total_gpa": {
                        "type": "number",
                        "title": "Total Gpa",
                        "default": 0.0
                    },
                    "total_credits": {
                        "type": "integer",
                        "title": "Total Credits",
                        "default": 0
                    },
                    "level": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StudentLevel"
                            }
                        ],
                        "default": "Freshman"
                    }
                },
                "type": "object",
                "required": [
                    "people_id",
                    "program",
                    "degree",
                    "curriculum",
                    "college",
                    "department"
                ],
                "title": "HDStudent"
            },
...
            "HDYear": {
                "properties": {
                    "year": {
                        "type": "integer",
                        "title": "Year"
                    },
                    "terms": {
                        "items": {
                            "$ref": "#/components/schemas/HDTerm"
                        },
                        "type": "array",
                        "title": "Terms",
                        "default": []
                    },
                    "year_gpa": {
                        "type": "number",
                        "title": "Year Gpa",
                        "default": 0.0
                    },
                    "total_credits": {
                        "type": "integer",
                        "title": "Total Credits",
                        "default": 0
                    }
                },
                "type": "object",
                "required": [
                    "year"
                ],
                "title": "HDYear"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RAResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "student_data": {
                        "$ref": "#/components/schemas/StudentData"
                    },
                    "graph_output": {
                        "type": "string",
                        "title": "Graph Output"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "student_data",
                    "graph_output"
                ],
                "title": "RAResponse"
            },
            "Skill": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "score"
                ],
                "title": "Skill"
            },
...
            "StudentLevel": {
                "type": "string",
                "enum": [
                    "Freshman",
                    "Sophomore",
                    "Junior",
                    "Senior"
                ],
                "title": "StudentLevel"
            },
            "StudentResponse": {
                "properties": {
                    "student": {
                        "$ref": "#/components/schemas/CompetenciesStudent"
                    },
                    "comparison_list": {
                        "items": {
                            "$ref": "#/components/schemas/CompetenciesStudent"
                        },
                        "type": "array",
                        "title": "Comparison List"
                    }
                },
                "type": "object",
                "required": [
                    "student",
                    "comparison_list"
                ],
                "title": "StudentResponse"
            },
...
            "XProfile": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "major": {
                        "type": "string",
                        "title": "Major",
                        "default": "Computer Science"
                    },
                    "department": {
                        "type": "string",
                        "title": "Department",
                        "default": "Computer Science Department"
                    },
                    "currentSemester": {
                        "type": "integer",
                        "title": "Currentsemester"
                    },
                    "enrolledCourses": {
                        "type": "integer",
                        "title": "Enrolledcourses"
                    },
                    "completedCourses": {
                        "type": "integer",
                        "title": "Completedcourses"
                    },
                    "coursesInProgress": {
                        "type": "integer",
                        "title": "Coursesinprogress"
                    },
                    "skills": {
                        "items": {
                            "$ref": "#/components/schemas/Skill"
                        },
                        "type": "array",
                        "title": "Skills"
                    },
                    "experiences": {
                        "items": {
                            "$ref": "#/components/schemas/Experience"
                        },
                        "type": "array",
                        "title": "Experiences"
                    },
                    "certificates": {
                        "items": {
                            "$ref": "#/components/schemas/Certificate"
                        },
                        "type": "array",
                        "title": "Certificates"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "name",
                    "password",
                    "currentSemester",
                    "enrolledCourses",
                    "completedCourses",
                    "coursesInProgress",
                    "skills",
                    "experiences",
                    "certificates"
                ],
                "title": "XProfile"
            }
        }
    }
}